Move types into their own files
