+++ libs/extension-messaging/src/types/message-listener.ts
@@ -1,0 +1,4 @@
+// The type for representing a listener that can be attached by any service.
+type MessageListener = (message: any) => Promise<any>;
+
+export default MessageListener;
