+++ libs/extension-messaging/src/types/message-result.ts
@@ -1,0 +1,7 @@
+// The internal result of processing a message.
+type MessageResult = {
+  success: boolean;
+  data: string;
+};
+
+export default MessageResult;
